refactor(app): dock markdown editor as right sidebar
